Add test cases for Mixpanel aggregate functions
