+++ packages/back-end/src/util/sql.ts
@@ -1,5 +1,4 @@
 import { format as sqlFormat } from "sql-formatter";
-import { Condition } from "../../types/metric";
 
 function getBaseIdType(objects: string[][], forcedBaseIdType?: string) {
   // If a specific id type is already chosen as the base, return it
@@ -114,42 +113,6 @@ export function format(sql: string) {
       // Similar fix for PrestoDB/Athena lambda functions
       .replace(/ - > /g, " -> ")
   );
-}
-
-export function getMixpanelPropertyColumn(col: string) {
-  // Use the column directly if it contains a reference to `event`
-  if (col.match(/\bevent\b/)) {
-    return col;
-  }
-
-  const colAccess = col.split(".").map((part) => {
-    if (part.substr(0, 1) !== "[") return `["${part}"]`;
-    return part;
-  });
-  return `event.properties${colAccess.join("")}`;
-}
-
-export function conditionToJavascript({ operator, value, column }: Condition) {
-  const col = getMixpanelPropertyColumn(column);
-  const encoded = JSON.stringify(value);
-
-  // Some operators map to special javascript syntax
-  if (operator === "~") {
-    return `(${col}||"").match(new RegExp(${encoded}))`;
-  } else if (operator === "!~") {
-    return `!(${col}||"").match(new RegExp(${encoded}))`;
-  } else if (operator === "=") {
-    return `${col}+'' == ${encoded}`;
-  } else if (operator === "=>") {
-    // Callback function
-    return `((value) => (${value}))(${col})`;
-  } else {
-    // If the value is a number, don't use the JSON encoded version for comparison
-    const comp = !value || isNaN(Number(value)) ? encoded : value;
-
-    // All the other operators exactly match the javascript syntax so we can use them directly
-    return `${col}+'' ${operator} ${comp}`;
-  }
 }
 
 // Recursively create list of metric denominators in order
